изменено:      Android/BluZ_LPM/app/src/main/java/ru/starline/bluz/BluetoothInterface.kt 	изменено:      Android/BluZ_LPM/app/src/main/java/ru/starline/bluz/drawCursor.kt 	изменено:      Android/BluZ_LPM/app/src/main/java/ru/starline/bluz/drawSpecter.kt 	изменено:      stm32cubeide/BluZ_LPM/Core/Inc/main.h 	изменено:      stm32cubeide/BluZ_LPM/Core/Src/main.c 	изменено:      stm32cubeide/BluZ_LPM/Release/BluZ_LPM.elf 	изменено:      stm32cubeide/BluZ_LPM/Release/BluZ_LPM.list 	изменено:      stm32cubeide/BluZ_LPM/Release/Core/Src/main.cyclo 	изменено:      stm32cubeide/BluZ_LPM/Release/Core/Src/main.o 	изменено:      stm32cubeide/BluZ_LPM/Release/Core/Src/main.su

+++ stm32cubeide/BluZ_LPM/Core/Inc/main.h
@@ -80,8 +80,8 @@ extern "C" {
 #define CHANNELS_2048	2048
 #define CHANNELS_4096	4096
 #define MAX_RESOLUTION 4096
-#define CAPCHAN 20
-#define LIMITCHAN 1 << CAPCHAN								/* Максимальное значение канала */
+#define CAPCHAN 20											/* Количество разрядов в канале */
+#define LIMITCHAN (uint32_t) (1 << CAPCHAN)					/* Максимальное значение канала */
 
 extern uint16_t MTUSizeValue;
 extern uint8_t resolution, dataType;
